+++ manager-ui/i18n/locales/fr_FR/groups.ts
@@ -160,6 +160,7 @@ export default {
           "invite-account": "Inviter un compte",
           "view-account-sessions": "Voir les sessions",
           "revoke-account-sessions": "Révoquer des sessions",
+          "purge-account-sessions": "Purger l'historique des sessions",
         },
         apiTokens: {
           access: "Accès",